Issue #84
At line 403 of the source code of bufr2ncObsTypes.py, instead of:

OutVals[VarSpec[0]] = BufrFloatToActual(Bval, VarSpec[2])

the code is updated:

OutVals[VarSpec[1]] = ...

The BUFR variable names are used instead of the NC variable names. At the def calc_obs_date_time -and not only- for the transformation of the time
the function will probably fail, everytime that somebody choose a different variable name for the nc files.